backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend service for a Pika-style AI video generation app."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "langchain>=0.3.0,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "langchain-google-genai>=2.1.0,<3.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "langgraph>=1.0.1",
    "python-dotenv>=1.2.2",
    "cryptography>=46.0.7",
    "google-auth>=2.49.2",
    "google-auth-oauthlib>=1.4.0",
    "google-auth-httplib2>=0.4.0",
    "google-api-python-client>=2.196.0",
    "sqlmodel>=0.0.38",
    "sqlalchemy[asyncio]>=2.0.49",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
]

[project.scripts]
backend = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
